backend/career-refinement-service/requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
json-repair>=0.25.0
supabase>=2.0.0
pydantic>=2.0.0
rich>=13.0.0
websockets>=13.0